featureflag: Add method to inject feature flag into context
To make it easier to inject "global" feature flags into a test context,
add a method that will inject a feature flag into both incoming and
outgoing contexts.

Also refactors the featureflag code a bit to add a helper
injectIntoIncomingAndOutgoingContext to DRY the general functionality of
adding a key into an incoming or outgoing context.