ci: Upgrade PgBouncer version to v1.16.1
We frequently see issues in CI where we're seemingly stuck terminating
connections to the database when using PgBouncer. While the root cause
is not clear at this point in time, PgBouncer has fixed a bug in v1.16.0
where connections were left hanging and didn't get cancelled as expected
when the pool size was full -- which may or may not be related to the
issues we're seeing.

Upgrade the PgBouncer version to v1.16.1 to potentially fix the issue.