ci: Introduce cache for Go modules
While we already have caches for build-time dependencies and Ruby Gems,
we don't have any for Go modules. As a result, each job needs to fetch
and build Go modules anew, which adds to the time it takes for these
jobs to finish.

Add a new Go module cache and include it as required.