ci: Drop reporting of Git version
In many of our jobs, we execute `git version` to report the version of
Git that was in use. In some of our jobs this is misleading though given
that we report the system-installed version of Git, which is not the one
that's used to execute our tests. And in other cases it's broken because
we execute the self-built version before calling `make git`, which only
works if we've been able to restore the cache containing this binary.

Remove reporting of the Git version altogether. We can derive which Git
version is in use by checking `GIT_VERSION` and our Makefile. While at
it, fix the logic which tries to delete our Git installation when
testing bundled Git in case the installation doesn't exist.