ref: Implement two-phase voting for DeleteRefs
The DeleteRefs RPC is voting on changes it's about to commit once after
all references have been locked, but it doesn't do a second vote to
confirm that deleting the refs has actually succeeded. Doing such
two-phase voting will soon become mandatory though with the introduction
of voting phases such that we can start to ignore errors if a
transaction failed to reach quorum in "prepared" state.

Implement support for two-phase voting in DeleteRefs to prepare for
this.

Changelog: changed