transaction: Introduce tracking manager
A lot of tests need to create a mock transaction manager to track the
transactional votes and assert that we've got the votes we expected.
This is quite repetitive and ultimately requires us to change a lot of
function signatures if the `transaction.Manager` interface changes.

Improve the situation by introducing a new `TrackingManager` which
records all votes.