replicator: Use ticker to process stale jobs
The function which processes stale jobs takes a configurable duration
for how often the monitoring loop should run. This makes it hard to
control the ticker from the outside, most importantly in tests.

Refactor the code to instead accept a ticker and amend tests to use a
manual ticker to get rid of timeouts.