featureflag: Enable atomic repository creation by default
With f87bc1e98 (Merge branch 'pks-create-repository-atomic' into
'master', 2021-11-24), we have implemented support for atomic repository
creation. With this mode, creation of repositories will use locking and
meaningful transactional voting to create repos in a race-free manner.

The feature flag guarding the new behaviour has been rolled out on
December 6th, and no issues have surfaced since then. But the change is
rather high-risk given that it changes semantics of some of the RPCs to
return errors in case the target repository exists already, which wasn't
the case previously. To be on the safe side, we thus want to keep this
feature flag for at least one release such that we have an escape hatch
in case there are issues with this change which we haven't yet seen.

Default-enable the flag.

Changelog: changed