cmd/praefect: replicate immediately after track-repository
track-repository is a subcommand that tracks a repository in the
praefect database. This is generally used by an admin when a repository
is on disk but not tracked in the praefect db. Once tracked, replication
to the other nodes takes about 5 minutes and requires the Praefect
process to be running. This MR adds a flag that, if set, will immediately
process a relication job and wait for it to finish.

Changelog: added