backup: Fix missing feature flag tests
The backup restoration tests aren't testing feature flags for atomic
creation and removal of repositories. This went undetected because the
tests used `context.Background()` to create the context instead of using
`testhelper.Context()`, which would've set up additional sanity checks
for feature flags.

Exercise the feature flags. This uncovers a bug in the test setup if
running with Praefect: because the repository entry does not exist in
the database, Praefect intercepted the repository removal and never
forwarded it to Gitaly. As a result, at the time we try to create the
repository the old repository still exists. This used to work alright,
but now fails with atomic repository creation enabled because the target
repository must not exist yet.

Fix this setup by creating the repository with `CreateRepository()`
first.