testhelper: Simplify feature set test
The test for `FeatureSet`s is needlessly complex, doing a lot of voodoo
with converting contexts. Simplify the test to instead just assert that
incoming and outgoing metadata are exactly the same such that we can
then proceed to only whether feature flags are set in one of both tests.