cache: Rework keyer tests to be independent
The keyer tests are all dependent on each other, and furthermore they're
written in terms of one long test. This makes it hard to see the
different parts we're trying to test, and furthermore it makes it harder
than necessary to refactor those tests.

Split up these tests into several subtests and have each of them use a
separate cache such that there can be no interdependencies.