featureflag: Fix setting incoming feature flags modifying parent context
When setting feature flags on a context, then the expectation is that it
will only be set for the child context which is returned from the
function. This doesn't work as expected though: the metadata we retrieve
from the current context is not a copy, so setting any new values on it
will modify the parent context, as well.

Fix this bug by copying the context before setting the key. Somehow,
this bug only seemed to impact incoming contexts: I wasn't able to
demonstrate that this is broken for outgoing contexts, even though they
should in theory suffer from the exact same problem. This commit fixes
the issue for both incoming and outgoing contexts regardless.

Changelog: fixed