featureflag: Demonstrate feature flags modifying parent context
When appending feature flags to a context, the expectation is that the
child context will have the feature flag set, but the parent context
will remain unmodified. This expectation is broken for incoming contexts
though, where setting a feature flag will set it in both contexts.

Add a test to demonstrate that this is broken.