Remove unused Project parameter from internal/allowed payload
Gitaly is sending the repository's absolute path to Rails via the
'project' parameter when it calls 'internal/allowed'. While Gitaly
is sending the absolute path of the repository, Rails and the protos
document this parameter as the full project path and explicitly say
this is not the disk path. This prompted me to look into it in more
details and it appears the parameter is not used. This commit drops the
parameter. We don't want to send disk paths to Rails as they won't be
what Rails expects them to be with Praefect rewriting relative paths.