datastore: Notification listener
The jackc/pgx driver differs in how it handles database
notifications send with NOTIFY statement. It requires us to
re-implement a listener of the notifications. The new
implementation does the same thing: it opens connection
and starts listening for the notifications on the channels,
The difference is that we are using a blocking method
to receive a notification instead of reading from the channel.
Also it doesn't provide low level access to the connection
state changes. That is why handler methods Connected() and
Disconnect() managed a bit differently, but the idea is the
same.