datastore: Listen for notifications resiliently
The Listen method continuously listens for the
notification until non-temporary error occurs.
Which means it will return once database connection
is interrupted/terminated/closed because of some
circumstances (network partitioning, database restart).
In that case the praefect won't be able to listen
for the notification anymore. That is why the new
ResilientListener type is added. It uses Listener
in the loop and attempts to always listen with some
await in case of an error.

We still support gitaly_praefect_notifications_reconnects_total
metric to track how many connection were established and
interrupted.