glsql: Support of the text arrays
We are using lib/pq.StringArray to deal with PostgreSQL
arrays of type TEXT. The pgx driver doesn't require a
special type to pass slices where arrays are expected,
but it still requires a special type to scan the value.
And this special type if pgtype.TextArray. But it is
not as simple as pq.StringArray and requires some additional
operations to get result as a slice of strings. That is
why glsql.StringArray type is created. It provides a
helper Slice() method that returns all non-null elements
from the array as a slice of strings.