testhelper: Add ability to create uncancellable contexts
In some contexts, we want to create test contexts which are not
cancellable to exercise certain behaviour in case it is missing. It's
still useful to inject feature flags in these cases though, so one
shouldn't just reach for `context.Background()` in such a case.

Introduce a new function `ContextWithoutCancel()` to address this
usecase.