praefect: Replace lib/pq with jackc/pgx
With all preparation work done in this commit the
lib/pq package is replaced by the jackc/pgx. As
it doesn't require a special types to be passed
to make an arrays we just skip wrapping.
At the same time we replace pq.StringArray with
glsql.StringArray to retrieve arrays of text.
And also we are forced to update expected error
messages in the tests.

Changelog: changed