lint: Disallow use of "normal" contexts
Our testhelper context has recently gained the ability to verify that
feature flags are exercised as expected. This is an important safety
guard which can help us to avoid releasing untested and thus potentially
buggy code.

Disallow usage of "normal" contexts which don't have this ability such
that we do not introduce their usage by accident.