cmd/praefect: create replication events
track-repository inserts database records for praefect to track the
repository, but doesn't insert any replication jobs. When a mutator RPC
comes through, it will trigger replication but we want to make sure that
no matter what the reositories are replicated to its secondaries.

Also refactors the tests to no longer test sql replication, as that is
no longer supported.

Changelog: fixed