cmd/praefect: add helper text to explain replication
Adds some text that explains what is happening with replication. The
command either creates replication jobs in the queue, or it replicates
immediately. In both of these cases, print out helpful messages.

This also adds a writer in the command so we don't rely on the logger so
the output is more reader friendly.

Changelog: added