Add GNU build-ids to each gitaly go binary
To more reliably support observability tools like perf and uprobes,
each generated binary needs to have its own unique gnu-build-id.

This implementation provides a GNU build-id that is derived from
the Go build-id.  This ensures uniqueness and also can inherit
determinism if the Go build-id provides it.

This is an alternative implementation to the simpler one in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4191

For more background on goals and trade-offs, see:
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1464#note_769013218