repository: Always enable locking RenameRepository RPC
We have been pushing our RPCs which manage repository locations to
always use locking semantics on the source repository and, if exists, on
the target repository. Like this, we cannot ever have two RPCs which try
to create and/or remove a repository location at the same point in time.
This allows us to properly use transactional voting for these RPCs.

In 0118edab0 (repository: Implement locking for RenameRepository,
2021-12-06), we have introduced locking support for the RenameRepository
RPC. This change has been successfull rolled out to production without
any issues on December 10th, so the change seems to work as intended.

Remove the feature flag and always enable locking semantics for
RenameRepository.

Changelog: changed