testdb: Rename file hosting logic to update health checks
The helper function to set up healthy nodes in the test database is
hosted in a file "db.go". This has been fine until now given that it was
the only function in this package anyway. But we're about to move the
logic to set up a test database and connect to it into this package,
too, at which point it becomes a bit awkward that it's hosted in such a
generic file.

Rename the file to "health.go" to prepare for the move.