rubyserver: Use `require.Eventually` to wait for Ruby server to come up
We must wait for the Ruby server to come up before we can exercise it in
its tests, which is implemented via a sleep-loop where we repeatedly
ping the server until it starts to serve requests. This pattern can be
implemented much more neatly with `require.Eventually()`. Refactor the
test to do so.