listenmux: Remove useless sleep
The listenmux package has a test which exercises that connections aren't
closed by gRPC if they're stolen by the muxer. One of these tests uses a
sleep to "give grpc-go a chance to close the connection", but in fact
the test reliably fails without the sleep if the handshaker is converted
to not steal the connection.

Remove the sleep.