Listen on localhost in tests
Some tests are opening a listener on all interfaces. The tests should
only listen on localhost and not open externally accessible ports.
On macOS, this also causes the firewall to prompt whether one wants to
allow connections to the port or not. This commit changes the said tests
to only listen on localhost.