nodes/tracker: Convert error tracker tests to not require sleeps
Error tracking uses an error window during which errors are considered.
Any error older than this window is simply ignored. To test this, we use
various sleeps in our tests such that we can be sure that the window is
over and that errors are treated as expected.

Convert the code to instead use an error window function which can be
injected by the caller. Like this, production code can use the normal
error window function which we currently use, and tests can use a
different one which doesn't use timing as the source and thus allows us
to get rid of the sleeps.