Fix TestRepo_SetConfig on macOS
Git sets extra config variables on macOS which TestRepo_SetConfig
isn't accounting for. This commit fixes the test on macOS by adding
the extra config keys to the set of expected entries.

The standardEntries slice was moved to the test body so the slice isn't
reused between the test cases. It worked by chance before as each of the
test cases had to allocate a new slice as the capacity of the standardEntries
was always full. With us now appending the macOS specific keys, the test
cases would overwrite each others entries.