gitaly: Inject the limit handler
The limit handler is currently created ad-hoc by Gitaly's server code.
We're about to change the way it hosts Prometheus metrics though such
that we can get rid of a set of global variables, and this will require
us to make sure that there's only a single instance of the limiter.

Inject the limit handler as a dependency when creating the server to
prepare for this.