limithandler: Move per-repository concurrency keyer
The limithandler supports use of different keys to limit concurrency by
via a keyer function. The main function used by production code is
hosted in an unrelated package, mostly because it's where the limiter is
currently initialized.

Move it into the limithandler package to make it available at a central
place.