license: Extend license find response with additional data
Currently, the work with licenses is split between Gitaly
and GitLab. GitLab uses Gitaly to figure out what is the
identifier of the license used in the project and then
uses it to retrieve other info related to the license.
It seems like doing the same work twice as the same library
is used to get identifier of the license and its info.
We extend response of the FindLicense RPC to include
additional detail about the license: name, URL and a path.
This should be enough to remove additional logic in the
GitLab.

The new go-based implementation uses another set of licenses
and requires additional changes to provide URL and name
of the license. It seems like we need to pay more attention
if new implementation suits our needs.