Fix TestMux_concurrency on macOS
TestMux_concurrency tests listemux' concurrency by running a lot of
concurrent connections. It also attempts to change the file limits of
the process. This doesn't really seem to be working on macOS as expected.
Running 100 concurrent connections causes issues as some connections get
reset. It should be sufficient to run the test with less concurrency, so
this commit drops the file limit changing code and sets the number of
concurrent connections to 50 from 100. This fixes the flakiness and 50
connections should be more than sufficient to assert the code works
concurrently.