praefect: Fix output of remove-repository
There was a bug in the logic so that we printed out "Did not remove"
even though we did remove the repository from disk. This change also
deletes some wrong code that checks for an error message that only
exists in praefect-but we are calling gitaly directly so this error
message will never appear.

Changelog: fixed