ssh: Convert to set hooks path via config
The ssh package overrides the hooks path by setting the global
`config.OverrideHooksPath` variable, which is deprecated and about to be
removed. Convert it to instead set up the hooks path via the Gitaly
configuration.

While at it, parallelize running tests now that we don't depend on
global state anymore.