wiki: Fix test setup to use hooks
The wiki service test setup is incomplete and doesn't set up Git hooks,
which is why we override the hooks path such that no hooks are ever
executed.

Fix the test setup to set up the hook service and build the gitaly-hooks
binary so that we can enable the use of hooks.