config: Introduce ability to override hooks path via config
Some tests cannot use Git hooks or want to inject their own hooks. Most
importantly, this includes tests which do not have a Gitaly service
around, and thus hooks wouldn't be able to connect to the hooks service.

Overriding hooks used to be implemented by either overriding the Ruby
directory, which implicitly also locates the hooks directory, or by
setting a global `hooks.Override` variable. While overriding the Ruby
directory works alright in most cases, it doesn't in cases where it also
needs to point to the real Ruby directory in case a Ruby server is
required for the test to function. And changing a global variable in
these cases is bad practice.

Introduce a new `HooksPath` variable into the Git configuration which
allows tests to override the path without either modifying the Ruby
directory or using the global variable. Use of the global variable will
be removed in subsequent commits.