gittest: Add commit option to write objects into alternate ODB
We have several tests which write commit objects into the alternate
object database. While some hand-code this, most use the helper function
`CreateCommitInAlternateObjectDirectory()`. This helper requires manual
assemlby of the command and is thus unnecessarily verbose to be used.

Add a new option to the `WriteCommit()` helper which allows callers to
specify an alternate ODB such that it can be used instead of the old
helper.