hook: Convert environment tests to be stricter
The tests which check whether environment variables are set up as
expected for custom hooks currently filter down the list of environment
variables to a specific subset. While this works, this is much too
lenient: we do not verify that unrelated environment variables get
filtered out in the hooks' environment, and furthermore we don't check
the value of other environment variables which aren't currently included
in this allowed-list.

Convert the test to instead use a deny-list of filtered variables such
that we only exclude a certain set of environment variables which we
know to be inconsistent across environments.