Disable global git configuration only in tests
We can't yet disable global git configuration as it would potentially
break setups which currently still rely on the global configuration.
We have to deprecate the global configs and wait until a major release
where we can drop support for them. We can already disable them in the
tests. While this is a dirty hack, it replaces a much more invasive hack
which was overwriting $HOME and breaking many setups. This commit can be
reverted once we are able to drop support for global configs completely.