gitaly/main: Add background goroutine to setup cgroups
In the previous commit we allow Gitaly startup to continue as normal
even if cgroups fail to setup. Most of the time this will be because on
a restart, the cgroup root hasn't been setup yet. This commit adds a
background task that continues to try to setup the cgroups on an
interval. This would allow the cgroup root to be setup after Gitaly
starts up, instead of requiring another restart for cgroups to come into
use.

Changelog: changed