operations: Simplify setup of custom hooks in tags test
Some tests which exercise UserCreateTag set up hooks which verify that
the new object is of the expected type when writing the reference. The
setup for these hooks is needlessly complicated, where we first write a
script into a temporary location, and then write another custom hook
script which executes the first script. This is done such that we can
easily inject the expected object type, but we can instead just write
the expected object type directly into the script itself.

Refactor the tests to do so and write the script as a custom hook
directly.