hook: Fix potential use of wrong Git in custom hooks
Custom hooks are executed via two different paths: either via a Git
command, or manually by directly calling into the hook manager. When
executing via a Git command, then Git will have known to adapt the PATH
environment variable such that its own path is prepended to it. The
result is that hook scripts will automatically pick up the correct Git
executable if they only call "git" without using a fully-qualified path.
This PATH variable gets picked up by gitaly-hooks and is forwarded to
the hook RPCs such that this mechanism works as expected.

We don't do this in the case where we directly call into the hook
manager though. Instead, we simply take PATH as it is defined in the
server's environment. As a result, it can happen that scripts wouldn't
pick up the correct Git executable, or fail to find it at all.

Fix this bug by prepending the Git executables' directory to PATH.

Changelog: fixed