hooks: Move generation of hooks path into Gitaly config
While the Gitaly config package stores the hook path, assembly of the
actual path as well as the ability to override the path are hosted in a
special "hooks" package. This is needlessly complex.

Move the code into Gitaly's "config" package and provide a new function
`HooksPath()` which computes the path.