danger: Remove tracking of team members
In our Dangerfile, we keep a list of team members, which is the perfect
kind of variable to grow stale. And of course it is stale given that
@jcaigitlab is missing from that list. Furthermore, its use is quite
limited: the only thing we use it for is to propose to add a
~group::gitaly label if the merge request author is part of the group.

Remove that group and the label-recommendation. While we could improve
this to just fetch the list of team members automagically, it doesn't
really feel worth.