Add versioning to Praefect disk storage path schemas
Praefect is about to begin generating unique relative paths for
repositories to avoid on disk conflicts. The current format of
`@repositories/xx/yy/<repo-id>` is sufficient for now and guarantees
unique storage directories for repositories. The uniqueness of the paths
is guaranteed by the repository ID sequence stored in Postgres. While
this is sufficient for now, it doesn't leave room for evolving the schema of
paths in the future. For example, if we wanted to ensure each replica lands
in a unique directory to avoid the same races we had with repositories
landing in the same directories, we'd have to define a new top level directory
in the storage to avoid conflicts with the existing paths. Having to define
new top level directories in the storage might be more confusing than having
a single Praefect owned directory that can contain the disk path changes.

This commit changes the top level directory in the storage from @repositories
to @praefect. This clearly marks the repository as Praefect owned. A version
number is added in to the path to allow for changing the expectations of the
uniquness of the directories. For example, when we in future make replicas land
in unique directories as well, we can do so by incrementing the path format
version and giving the replicas paths like `@praefect/v2/xx/yy/<replica_id>`.
If we didn't increment the version, the replica ids could conflict with the
repository ids in the current path schema.

Praefect is not yet generating unique paths, so this is a good time to change
the format before creating the @repositories directory in the storages. Nothing
would break even if we did but we can directly go to a format that can handle
evolving the schema in the future.