git: Expose execution environment via command factory
The execution environment of Git, that is both where the Git executable
resides and the set of environment variables required such that Git can
find auxiliary Git commands, is currently hosted in Gitaly's config
code. As a result, many sites which need to somehow set up a Git call
reach into the configuration's fields directly. This has worked until
now, but as it turns out it is really hard to refactor execution paths
like this. Most importantly, we cannot easily introduce feature flags
for how Git itself is executed.

To fix this and be able to easily use feature flags we need to move this
functionality into a place which can handle feature flags. The most
natural place would thus of course be the Git commmand factory, which
would thus become the single source of truth about where the Git binary
resides and the required environment variables.

Expose a new function to retrieve the execution environment to the
factory's interface. This is a first step only and allows us to migrate
all existing callsites which reach into the configuration right now to
instead ask the command factory.