rubyserver: Resolve Git binary path via command factory
The Git binary path for the gitaly-git2go executor is currently resolved
by accessing the configured Git path directly. This is deprecated, where
callers should instead ask the Git command factory for the Git execution
environment.

Convert the Ruby server to accept a Git command factory and query it for
the binary path. Unfortunately, we cannot use a proper context here
given that the Git configuration is passed in at creation time of the
server. This essentially means that use of feature flags to control the
Git execution environment doesn't work in the context of the Ruby
server.