metrics: Total amount of connections to the service
The existing metric 'gitaly_connections_total' exists
on a package level (the global metric) and used for
both praefect and gitaly services. With this change
we address both those problems: the metric now are
local to the service and a new metric introduced
specially for praefect following the same naming
pattern as all other praefect-related metrics.

The content of the connectioncounter package moved
into a starter package as it makes no sense to have
it separately.

Change: added