gittest: Simplify execution of Git commands
Refactor the gittest helper functions to execute Git commands to be
easier to extend. It reimplements `Exec()` to just call `ExecOpts()`,
passes in the `ExecConfig` into `run()` and drops the special-case where
the caller passed in a stdout writer, which wasn't needed at all. These
refactorings will allow us to more easily introduce a new helper
function which preconstructs the `exec.Cmd`, but doesn't execute it yet.