gittest: Filter environment when running Git commands
When configuring Git in our testhelper, we execute git-rev-parse(1) to
find all environment variables which may impact the execution of Git and
unset them. This is done to protect us against inadvertently using or
modifying for example the wrong Git repository if GIT_DIR was set. While
it works, unsetting the environment variables keeps us from actually
exercising the code that should prevent such things to happen via
`command.AllowedEnvironment()`, which should always be used to filter
down variables to a known-good set.

Drop the code which unsets these environment variables and fix callsites
which failed to filter environment variables correctly. Next to the
gittest package, this also impacts the Ruby server which didn't sanitize
its environment variables correctly. As a result, a misconfigured
environment could've led to all kinds of misbehaviour in the Ruby server
and its spawned Git commands.

This change may lead to some overfiltering of environment variables in
the Ruby server though, which may impact its ability to find its Ruby
Gems. This should hopefully be fixed by manually allowing the `GEM_HOME`
envvar.