gittest: Add sanity check whether Git configuration is ignored
When spawning Git commands, we should always sanitize the environment
variables such that only known-good values are allowed which don't
impact the running Git command. Two such environment variables are
GIT_CONFIG_GLOBAL and GIT_CONFIG_SYSTEM, which can be used to set up
alternate Git configuration files for the global and system scope. If
these are set, then we should ignore them as otherwise we may not be
running Git commands with the expected Git configuration.

Set those environment variables when executing tests such that we can
assert that they get stripped when running Git commands. As it turns
out, there is one more location which failed to do the filtering
correctly, which is the gitaly-hooks command.

While at it, move setup of those environment variables from our Makefile
into our test configuration such that it's used even if not executing
tests via the Makefile.