testserver: Extract function to start Praefect server
The testserver package contains logic to both start Gitaly and Praefect
servers from a given configuration, but the latter is only used to start
a Praefect proxy for the "test-with-praefect" target. It is more
generally useful though in other locations where we start Praefect.

Refactor the code and pull the logic to start Praefect into a separate
`StartPraefect()` server such that it can be used from outside of the
package.