hook: Let custom hook scripts resolve Git via PATH
Custom hook scripts used in tests which use Git currently have the fully
qualified Git path injected into them such that we are sure that the
correct Git version is used. This was required until now given that we
didn't inject a PATH environment variable which had the correct Git
version prepended to it. As a result, we'd have executed the wrong Git
executables in these scripts, which would typically be our intercepting
Git wrappers in `internal/testhelper/testdata/home/bin`.

Now that we do properly set up PATH, this workaround isn't needed
anymore. Quite the reverse: the fact that we injected the full Git path
into these scripts only obscured the fact that we didn't set up the
environment as expected.

Adapt tests to always resolve Git via PATH to assert that we pick up the
correct executables.